config/micro-config/AIplugin/requirements.txt
google-generativeai
mako
textual
textual-dev
